… zio-json/shared/src/main/scala/zio/json/internal/lexer.scala