Tested with Scala 3 and JDK-21 on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                              (size)   Mode  Cnt       Score       Error  Units
ArrayOfDurationsReading.zioJson           128  thrpt    5  139635.386 ±  6279.622  ops/s
ArrayOfInstantsReading.zioJson            128  thrpt    5  145895.788 ±  2712.227  ops/s
ArrayOfLocalDateTimesReading.zioJson      128  thrpt    5  145501.719 ±   954.572  ops/s
ArrayOfLocalDatesReading.zioJson          128  thrpt    5  256254.145 ±  4308.849  ops/s
ArrayOfLocalTimesReading.zioJson          128  thrpt    5  204250.156 ±  7884.575  ops/s
ArrayOfOffsetDateTimesReading.zioJson     128  thrpt    5  120077.595 ±  2159.185  ops/s
ArrayOfOffsetTimesReading.zioJson         128  thrpt    5  163593.607 ±  5666.016  ops/s
ArrayOfPeriodsReading.zioJson             128  thrpt    5  132556.463 ±  4627.411  ops/s
ArrayOfZoneIdsReading.zioJson             128  thrpt    5  194228.482 ±  9299.752  ops/s
ArrayOfZoneOffsetsReading.zioJson         128  thrpt    5  293216.532 ±  4029.065  ops/s
ArrayOfZonedDateTimesReading.zioJson      128  thrpt    5   58358.983 ±  1603.588  ops/s
```

#### After
```txt
Benchmark                              (size)   Mode  Cnt       Score       Error  Units
ArrayOfDurationsReading.zioJson           128  thrpt    5  156375.366 ± 27386.346  ops/s
ArrayOfInstantsReading.zioJson            128  thrpt    5  162087.028 ±   846.403  ops/s
ArrayOfLocalDateTimesReading.zioJson      128  thrpt    5  192870.823 ±  5247.402  ops/s
ArrayOfLocalDatesReading.zioJson          128  thrpt    5  328219.618 ±  5653.266  ops/s
ArrayOfLocalTimesReading.zioJson          128  thrpt    5  263869.206 ±  7690.763  ops/s
ArrayOfOffsetDateTimesReading.zioJson     128  thrpt    5  145568.759 ±  2617.131  ops/s
ArrayOfOffsetTimesReading.zioJson         128  thrpt    5  190794.832 ±  1147.860  ops/s
ArrayOfPeriodsReading.zioJson             128  thrpt    5  160170.830 ± 20311.599  ops/s
ArrayOfZoneIdsReading.zioJson             128  thrpt    5  195976.833 ±  9716.529  ops/s
ArrayOfZoneOffsetsReading.zioJson         128  thrpt    5  423405.782 ±  8617.047  ops/s
ArrayOfZonedDateTimesReading.zioJson      128  thrpt    5   70711.431 ±  2026.515  ops/s
```